+++ CHANGELOG.md
@@ -11,11 +11,12 @@ Anticipated release: July 9, 2021
 
 #### ⚙️ Behind the scenes
 
+- add tests for authentication ([#3188])
 - install cypress for end-to-end testing ([#3226])
-
 # Previous releases
 
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
+[#3188]: https://github.com/CMSgov/eAPD/issues/3188
 [#3226]: https://github.com/CMSgov/eAPD/issues/3226
 [#2996]: https://github.com/CMSgov/eAPD/issues/2996